+++ Sources/Client/Utils/Atomic.swift
@@ -9,14 +9,17 @@
 import Foundation
 
 /// A mutable thread safe variable.
+///
+/// - Note: Even though the value guarded by `Atomic` is thread-safe, the `Atomic` class itself is not. Mutating the instance
+/// itself from multiple threads can cause a crash.
 @dynamicMemberLookup
 public final class Atomic<T> {
     /// A didSet callback type.
-    public typealias DidSetCallback = (_ value: T?, _ oldValue: T?) -> Void
+    public typealias DidSetCallback = (_ value: T, _ oldValue: T) -> Void
 
     private let lock = NSRecursiveLock()
 
-    private var value: T? {
+    private var value: T {
         didSet {
             if let callbackQueue = callbackQueue {
                 callbackQueue.async {
@@ -32,25 +35,26 @@ public final class Atomic<T> {
     private var didSet: DidSetCallback?
     private var callbackQueue: DispatchQueue?
 
-    /// Init a Atomic.
+    /// Creates a new `Atomic` instance.
     ///
     /// - Parameters:
-    ///   - value: an initial value.
-    ///   - didSet: a didSet callback.
-    public init(_ value: T? = nil, callbackQueue: DispatchQueue? = .global(qos: .userInitiated), _ didSet: DidSetCallback? = nil) {
+    ///   - value: The initial value.
+    ///   - callbackQueue: The queue which is used for `didSet` callback calls.
+    ///   - didSet: Called after the current value of `Atomic` is changed.
+    public init(_ value: T, callbackQueue: DispatchQueue? = .global(qos: .userInitiated), _ didSet: DidSetCallback? = nil) {
         self.value = value
         self.callbackQueue = callbackQueue
         self.didSet = didSet
     }
     
     /// Set a value.
-    public func set(_ newValue: T?) {
+    public func set(_ newValue: T) {
         lock.lock()
         value = newValue
         lock.unlock()
     }
 
-    public func get() -> T? {
+    public func get() -> T {
         lock.lock(); defer { lock.unlock() }
         return value
 
@@ -60,17 +64,16 @@ public final class Atomic<T> {
     ///
     /// - Parameter default: a default value.
     /// - Returns: a stored value or default.
+    @available (*, deprecated, message: "Using `get(default:)` for non-optional types is deprecated because it has no effect.")
     public func get(default: T) -> T {
-        get() ?? `default`
+        get()
     }
     
     /// Update the value safely.
     /// - Parameter changes: a block with changes. It should return a new value.
-    public func update(_ changes: (T) -> T?) {
+    public func update(_ changes: (T) -> T) {
         lock.lock()
-        if let currentValue = value {
-            value = changes(currentValue)
-        }
+        value = changes(value)
         lock.unlock()
     }
 }
@@ -104,17 +107,44 @@ public extension Atomic {
     }
     
     /// Accesses a sub value by the given keypath.
-    subscript<Element>(dynamicMember keyPath: WritableKeyPath<T, Element>) -> Element? {
+    subscript<Element>(dynamicMember keyPath: WritableKeyPath<T, Element>) -> Element {
         get {
-            return get()?[keyPath: keyPath]
+            return get()[keyPath: keyPath]
         }
         set {
-            if let newValue = newValue {
-                update(keyPath, to: newValue)
-            }
+            update(keyPath, to: newValue)
         }
     }
 }
+
+// MARK: - Helpers for optional T
+
+// swiftlint:disable syntactic_sugar
+extension Atomic {
+    
+    /// Creates a new `Atomic` instance with the initial value of `nil`.
+    ///
+    /// - Parameters:
+    ///   - callbackQueue: The queue which is used for `didSet` callback calls.
+    ///   - didSet: Called after the current value of `Atomic` is changed.
+    public convenience init<Wrapped>(callbackQueue: DispatchQueue? = .global(qos: .userInitiated),
+                                     _ didSet: DidSetCallback? = nil) where T == Optional<Wrapped> {
+        self.init(.none, callbackQueue: callbackQueue, didSet)
+    }
+    
+    /// Returns the current value if not `nil` or returns the default value.
+    ///
+    /// - Parameter default: The value used if the current value of `Atomic` is `nil`.
+    public func get<Wrapped>(default: Wrapped) -> Wrapped where T == Optional<Wrapped> {
+        switch get() {
+        case .none:
+            return `default`
+        case let .some(some):
+            return some
+        }
+    }
+}
+// swiftlint:enable syntactic_sugar
 
 // MARK: - Helper for Collection
 
@@ -123,7 +153,7 @@ public extension Atomic where T: Collection {
     // swiftlint:disable:next syntactic_sugar
     subscript<Key: Hashable, Value>(key: Key) -> Value? where T == Dictionary<Key, Value> {
         lock.lock(); defer { lock.unlock() }
-        return value?[key]
+        return value[key]
     }
 }